Implement AVX2 optimized matmul()
The matmul() function is really a simplified sgemv rather than the sgemm
that you would expect. This version gives us about 30% faster token
generation on a Ryzen 7 5800X and outperforms both Intel and OpenBLAS'
cblas_sgemv() functions. It relies on the CPU's prefetch rather than
manually prefetching and the main innovation that enables its high
performance is an extremely efficient horizontal addition.

Instead of using 5 AVX2 instructions for horizontal addition or
switching to scalar arithmetic, it does 11 AVX2 instructions to perform
8 horizontal additions in parallel. The end result is a single vector
containing the results of 8 horizontal additions that we can then write
to main memory directly.

The function assumes 32-byte memory alignment, which the earlier
calloc_aligned() patch gives us. It has handling for non multiple of 8
matrix/vector dimensions, but in practice, all matrices and vectors
passed to it by llama3.c have dimensons that are multiples of 8, so the
handling is unnecessary and is written in plain C rather than
intrinsics, since doing performance work on effectively dead code gives
zero benefit. The code to handle non-powers of 8 is only included for
completeness, although it is untested.

Signed-off-by: Richard Yao <richard.yao@alumni.stonybrook.edu>