Use cblas_sgemm_batch() for attention calculations
Note that I am not actually sure why the second cblas_sgemm_batch()
works as gemm shouldn't be able to calculate what is being calculated,
but the output is the same in testing and performance increases
substantially, so I am going with it for now. I will research it more
later.

Signed-off-by: Richard Yao <richard.yao@alumni.stonybrook.edu>