Replace part of attention loop in prompt processing using SGEMM
Rather than multiplying a vector by a matrix, for each position/head
combination, we can multiply a matrix by a matrix for each head. This is
much faster because it is not memory bandwidth bound.

This leaves some room on the table for future optimization as we are
technically multiplying a rectangular matrix by a matrix. About half of
the calculations are unnecessary.

Signed-off-by: Richard Yao <richard.yao@alumni.stonybrook.edu>