rung.c: Remove cudaMemcpyAsync() from forward()
I read that CUDA calls are asynchronous, but are executed serially
within a stream. I already read that cudaMemcpy() is a synchronization
point. This means that copying asynchronously in a separate stream could
allow a subsequent iteration of the forward function to overwrite the
array before it has been used, which is undefined behavior that would
cause incorrect calculation results. Until I am sure about the safety, I
am reverting this. The use of the asynchronous function in a separate
cuda stream did not make a noticeably difference in performance in the
first place, so this difference probably will not be noticeable in any
benchmark.

Signed-off-by: Richard Yao <richard.yao@alumni.stonybrook.edu>